apps/backend/requirements.txt
# LLM & Orchestration
langchain
langchain-google-genai
langchain-aws
langchain-groq
google-genai
boto3
pydantic

# Math & Logic
numpy

# Database
pyodbc

# Environment & API
python-dotenv
fastapi
uvicorn
